scorescope ai/Exam_analytics_hackathon/Exam_analytics_Hack/exam2/backend/accounts/views.py
from rest_framework import generics, permissions, status
from rest_framework.response import Response
from rest_framework.views import APIView
from django.contrib.auth import get_user_model
from .serializers import RegisterSerializer, UserSerializer
from analytics.hindsight_service import hindsight_service

User = get_user_model()


class RegisterView(generics.CreateAPIView):
    """Register a new student account."""
    queryset = User.objects.all()
    serializer_class = RegisterSerializer
    permission_classes = [permissions.AllowAny]

    def create(self, request, *args, **kwargs):
        serializer = self.get_serializer(data=request.data)
        if not serializer.is_valid():
            # Flatten all errors into a single readable message
            errors = {}
            for field, messages in serializer.errors.items():
                errors[field] = messages[0] if isinstance(messages, list) else str(messages)
            return Response({
                'errors': errors,
                'message': list(errors.values())[0] if errors else 'Registration failed',
            }, status=status.HTTP_400_BAD_REQUEST)

        user = serializer.save()
        # Best-effort: retain user preferences in Hindsight.
        try:
            hindsight_service.retain_user_profile(user)
        except Exception:
            pass
        return Response({
            'user': UserSerializer(user).data,
            'message': 'Registration successful'
        }, status=status.HTTP_201_CREATED)


class ProfileView(APIView):
    """Get or update current user's profile."""

    def get(self, request):
        serializer = UserSerializer(request.user)
        return Response(serializer.data)

    def put(self, request):
        serializer = UserSerializer(request.user, data=request.data, partial=True)
        serializer.is_valid(raise_exception=True)
        user = serializer.save()
        # Best-effort: update retained user preferences.
        try:
            hindsight_service.retain_user_profile(user)
        except Exception:
            pass
        return Response(UserSerializer(user).data)
