scorescope ai/Exam_analytics_hackathon/Exam_analytics_Hack/exam2/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
scikit-learn>=1.3
pandas>=2.1
numpy>=1.24
hindsight-client
groq>=0.4
python-dotenv>=1.0
